refactor(codec): split numeric binary codecs
